Combine & Renormalize

Combines samples, rescales and renormalizes, runs Dimensional Reduction, and returns a combined Seurat Object. This template will summarize the multi-dimensionality of your data into a set of "principal components" to allow for easier analysis. There is an option to use this template to perform Integration, as well. This is Step 4 in the canonical Single Cell pipeline after elbow plots and regression.

Combine_and_Renormalize(
  Seurat_Object,
  npcs = 15,
  vars_to_regress = NULL,
  integratedata = FALSE,
  clust_res_low = 0.2,
  clust_res_high = 1.2,
  clust_res_bin = 0.2,
  only_var_genes = FALSE,
  draw_umap = TRUE,
  draw_tsne = TRUE,
  imageType = "png",
  nfeatures = 2000,
  low_cut = 0.1,
  high_cut = 8,
  low_cut_disp = 1,
  high_cut_disp = 1e+05,
  selection_method = "vst",
  cell_hashing_data = FALSE,
  project_name = "scRNAProject",
  doMergeData = TRUE,
  seed_for_PCA = 42,
  seed_for_TSNE = 1,
  seed_for_UMAP = 42,
  Do_SCTransform = TRUE,
  Exclude_sample = 0
)

Arguments

Seurat_Object

description.

npcs

Select the number of principal components for your analysis. Please see the elbow plot in the previous template to figure out what number of PCs explains your variance cut-off. For example, if the elbow plot has point at (15,0.02), it means that 15 PCs encapsulate 98% of the variance in your data.

vars_to_regress

Subtract (‘regress out’) this source of heterogeneity from the data. For example, to regress out mitochondrial effects, input "percent.mt." Options: percent.mt, nCount_RNA, S.Score, G2M.Score, CC.Difference

integratedata

Perform integration of cells across conditions using the most variant genes to identify cells most similar to each other. N.B. Always look at cells before deciding whether to perform integration, and seek advice from bioinformatician.

clust_res_low

Select minimum resolution for clustering plots. The lower you set this, the FEWER clusters will be generated.

clust_res_high

Select the maximum resolution for clustering. The higher you set this number, the MORE clusters you will produced.

clust_res_bin

Select the bins for your cluster plots. For example, if you input 0.2 as your bin, and have low/high resolution ranges of 0.2 and 0.6, then the template will produce cluster plots at resolutions of 0.2, 0.4 and 0.6.

only_var_genes

If dataset is larger than ~40k filtered cells, toggle to TRUE. If TRUE, only variable genes will be available for downstream analysis. Default is FALSE.

draw_umap

If TRUE, draw UMAP plot.

draw_tsne

If TRUE, draw TSNE plot.

imageType

Select output image type. Options: png, svg

nfeatures

Number of variable features.

low_cut

description.

high_cut

description.

low_cut_disp

description.

high_cut_disp

description.

selection_method

Method to choose top variable features. Options: vst, mean.var.plot, dispersion

cell_hashing_data

Toggle "true" if you are using cell-hashed data.

project_name

description.

doMergeData

Toggle FALSE to stop metadata from merging. Not recommended for standard pipeline.

seed_for_PCA

description.

seed_for_TSNE

description.

seed_for_UMAP

description.

Do_SCTransform

Set to TRUE to run SCTransform (recommended current v3 Seurat default). Set to FALSE to run ScaleData (previous v2 Seurat default) instead.

Exclude_sample

Exclude unwanted samples from the merge step. The number will correspond to the order in which they appear. Leave as 0 if you want to use all samples. If you want to exclude one or several samples, separate each sample number by comma (e.g. 1,2,3,4).

Value

Seurat Objects and QC plots. Combines samples, rescales and renormalizes, runs Dimensional Reduction, and returns a combined Seurat Object. This template will summarize the multi-dimensionality of your data into a set of "principal components" to allow for easier analysis. There is an option to use this template to perform Integration, as well. This is Step 4 in the canonical Single Cell pipeline after elbow plots and regression.

Details

Combine_and_Renormalize template in single-cell-rna-seq-r4 NIDAP environment from v178