Combine & Renormalize

Combines samples, rescales and renormalizes, runs Dimensional Reduction, and returns a combined Seurat Object. This template will summarize the multi-dimensionality of your data into a set of "principal components" to allow for easier analysis. There is an option to use this template to perform Integration, as well. This is Step 4 in the canonical Single Cell pipeline after elbow plots and regression.

Combine_and_Renormalize_DEV(
  Seurat.Object,
  npcs = 15,
  vars.to.regress = NULL,
  integratedata = FALSE,
  clust.res.low = 0.2,
  clust.res.high = 1.2,
  clust.res.bin = 0.2,
  only.var.genes = FALSE,
  draw.umap = TRUE,
  draw.tsne = TRUE,
  imageType = "png",
  nfeatures = 2000,
  low.cut = 0.1,
  high.cut = 8,
  low.cut.disp = 1,
  high.cut.disp = 1e+05,
  selection.method = "vst",
  cell.hashing.data = FALSE,
  project.name = "scRNAProject",
  seed.for.PCA = 42,
  seed.for.TSNE = 1,
  seed.for.UMAP = 42,
  SCTransform = TRUE,
  exclude.sample = ""
)

Arguments

npcs

Select the number of principal components for your analysis. Please see the elbow plot in the previous template to figure out what number of PCs explains your variance cut-off. For example, if the elbow plot has point at (15,0.02), it means that 15 PCs encapsulate 98% of the variance in your data.

integratedata

Perform integration of cells across conditions using the most variant genes to identify cells most similar to each other. N.B. Always look at cells before deciding whether to perform integration, and seek advice from bioinformatician.

clust.res.low

Select minimum resolution for clustering plots. The lower you set this, the FEWER clusters will be generated.

clust.res.high

Select the maximum resolution for clustering. The higher you set this number, the MORE clusters you will produced.

clust.res.bin

Select the bins for your cluster plots. For example, if you input 0.2 as your bin, and have low/high resolution ranges of 0.2 and 0.6, then the template will produce cluster plots at resolutions of 0.2, 0.4 and 0.6.

only.var.genes

If dataset is larger than ~40k filtered cells, toggle to TRUE. If TRUE, only variable genes will be available for downstream analysis. Default is FALSE.

draw.umap

If TRUE, draw UMAP plot.

draw.tsne

If TRUE, draw TSNE plot.

imageType

Select output image type. Options: png, svg

nfeatures

Number of variable features.

low.cut

description.

high.cut

description.

low.cut.disp

description.

high.cut.disp

description.

selection.method

Method to choose top variable features. Options: vst, mean.var.plot, dispersion

cell.hashing.data

Toggle "true" if you are using cell-hashed data.

project.name

description.

seed.for.PCA

description.

seed.for.TSNE

description.

seed.for.UMAP

description.

SCTransform

Set to TRUE to run SCTransform (recommended current v3 Seurat default). Set to FALSE to run ScaleData (previous v2 Seurat default) instead.

exclude.sample

Exclude unwanted samples from the merge step. Include sample names to be removed. If you want to exclude several samples, separate each sample number by comma (e.g. sample1,sample2,sample3,sample4).

object

description.

vars_to_regress

Subtract (‘regress out’) this source of heterogeneity from the data. For example, to regress out mitochondrial effects, input "percent.mt." Options: percent.mt, nCount_RNA, S.Score, G2M.Score, CC.Difference

Value

Seurat Objects and QC plots. Combines samples, rescales and renormalizes, runs Dimensional Reduction, and returns a combined Seurat Object. This template will summarize the multi-dimensionality of your data into a set of "principal components" to allow for easier analysis. There is an option to use this template to perform Integration, as well. This is Step 4 in the canonical Single Cell pipeline after elbow plots and regression.

Details

Combine_and_Renormalize template in single-cell-rna-seq-r4 NIDAP environment from v178