Post-filter QC Plots

This set of plots compares several QC parameters across samples after initial QC and filtering. This template is Step 2 in the basic Single-Cell RNA-seq workflow.

Post_filter_QC(Seurat_Object, Parallelize_Computation = F, Image_type = "png")

Arguments

Seurat_Object

Please input a filtered Seurat Object. This should be the output from the Initial QC template.

Parallelize_Computation

Toggle true to parallelize your computations using Spark. Recommended for large datasets.

Image_type

Remember that svgs are much larger than pngs, so we recommend doing everything first in png, then rerunning to output specific svgs as needed.

Value

Seurat Objects and QC plots that compares several QC parameters across samples after initial QC and filtering. This template is Step 2 in the basic Single-Cell RNA-seq workflow.

Details

Post-filter QC Plots template in single-cell-rna-seq-r4 NIDAP environment from v47