Annotating cell types using SingleR module

SingleR is an automatic annotation method for single-cell RNA sequencing (scRNAseq) data (Aran et al. 2019). Given a reference dataset of samples (single-cell or bulk) with known labels, it labels new cells from a test dataset based on similarity to the reference.

Usage

annotateCellTypes(
  object,
  species = "Mouse",
  reduction.type = "umap",
  legend.dot.size = 2,
  do.finetuning = FALSE,
  local.celldex = NULL,
  use.clusters = NULL
)

Arguments

object

Object of class Seurat (your combined Seurat Object after PC reduction has been performed)

species

The species of your samples ("Human" or "Mouse"). Default is "Mouse"

reduction.type

Select the kind of clustering visualization you would like to use to visualize the cell type results ("umap", "tsne", "pca")

legend.dot.size

the size of the colored dots on your chart legend. Default is 2

do.finetuning

Performs the SingleR fine-tuning function. Default is FALSE

local.celldex

Provide a local copy of CellDex library. Default is NULL

use.clusters

Provide cluster identities for each cell. Default is NULL

Value

a Seurat object with additional metadata

Details

This function is Step 5 of the basic Single-Cell RNA-seq workflow. It is the starting point for downstream visualization, subsetting, and analysis. It takes a combined seurat object as input, such as the one created by the Combined&Renormalized function at the end of the Filter&QC Path