Color by Gene List

Returns a panel of reduction plots colored by marker expression

Usage

colorByMarkerTable(
  object,
  samples.subset,
  samples.to.display,
  manual.genes = c(),
  marker.table,
  cells.of.interest,
  protein.presence = FALSE,
  assay = "SCT",
  slot = "scale.data",
  reduction.type = "umap",
  point.transparency = 0.5,
  point.shape = 16,
  cite.seq = FALSE
)

Arguments

object

Seurat-class object

samples.subset

List of samples to subset the data by

samples.to.display

List of samples to depict on dimension plot, samples not in the list would be colored gray in the background

marker.table

Table of marker genes for each celltype (column names of the table), append "_prot" or "_neg" for proteins or negative markers

cells.of.interest

Celltypes from geneset_dataframe to screen for

protein.presence

Set to TRUE if protein markers are used

assay

Assay to extract gene expression data from (Default: "SCT")

reduction.type

Choose among tsne, umap, and pca (Default: "umap")

point.transparency

Set to lower values for more see through points on dimension plot (Default: 0.5)

point.shape

Change the shape of points for between visualization (Default: 16)

cite.seq

Set to TRUE to use CITE-seq embedding for dimension reduction

Value

arranged grob of dimension reduction plots colored by individual marker expression

Details

Takes in a gene table inputted by the user, displays a panel of tsne, umap, or pca colored by marker expression. The panel will be organized in a similar format as the gene table, but with the omission of genes not found in the data