DEG (Gene Expression Markers)

This function performs a DEG (differential expression of genes) analysis on a merged Seurat object to identify expression markers between different groups of cells (contrasts).

Usage

degGeneExpressionMarkers(
  object,
  samples,
  contrasts,
  parameter.to.test = "orig_ident",
  test.to.use = "MAST",
  log.fc.threshold = 0.25,
  use.spark = FALSE,
  assay.to.use = "SCT"
)

Arguments

object

Seurat-class object

samples

Samples to be included in the analysis

contrasts

Contrasts in the "A-B" format

parameter.to.test

Select the metadata column that you would like to use to perform your DEG analysis and construct your contrasts from. Default is "orig_ident"

test.to.use

The kind of algorithm you would like to use to perform your DEG analysis. Default is the MAST algorithm (wilcox,bimod,roc,t,negbinom,poisson,LR,MAST,DESeq2).

log.fc.threshold

The minimum log fold-change between contrasts that you would like to analyze. Default is 0.25

use.spark

Opt to use Spark to parallelize computations. Default is FALSE

assay.to.use

The assay to use for your DEG analysis. Default is SCT, but can use linearly scaled data by selecting RNA instead

Value

a dataframe with DEG.

Details

The recommended input is a merged Seurat object with SingleR annotations, along with its associated sample names and metadata