Dotplot of Gene Expression by Metadata

This function uses the Dotplot function from Seurat and plots average gene expression values and percent expressed for a set of genes.

Usage

dotPlotMet(
  object,
  metadata,
  cells,
  markers,
  use_assay = "SCT",
  plot.reverse = FALSE,
  cell.reverse.sort = FALSE,
  dot.color = "darkblue"
)

Arguments

object

Seurat Object

metadata

Metadata column in Seurat Object to plot

cells

Vector of metadata category factors to plot and should be found in metadata column. Order of plotting will follow exact order as entered.

markers

Vector of genes to plot. Order of plotting will follow exact order as entered

plot.reverse

If TRUE, set metadata categories to x-axis and genes to y-axis (default is FALSE)

cell.reverse.sort

If TRUE, Reverse plot order of metadata category factors (default is FALSE)

dot.color

Dot color (default is "dark blue")

Value

Dotplot with markers and cell types.

Details

This method provides a dotplot showing the percent frequency of gene-positive cells as size of dot and degree of expression as color of dot.