Pseudobulk Differential Gene Expression Analysis

Performs Linear Models for Microarray Data on single-cell data

pseudobulkDEG(
  so,
  contrasts,
  replicate,
  comparison.level,
  label,
  min.num.cells = 3,
  min.num.reps = 1,
  min.num.features = 0,
  subdivide_data = FALSE
)

Arguments

so

Seurat-class object

contrasts

Choose the groups (group1-group2) to compare transcription profiles against, positive Fold Change values indicate upregulation in group1

replicate

Column of metadata containing replicate information, gene expression will first be aggregated based on this column before differential analysis is ran

comparison.level

Name of metadata column containing information on the sub-levels you want to divide your data into.

label

Name of metadata column containing information on the groups (contrasts) you want to compare.

min.num.cells

Filters a comparison level based on the minimum number of cells at that level.

min.num.reps

Filters a comparison level based on the minimum number of replicates at that level.

min.num.features

Filters out genes if they are expressed in fewer than X number of cells.

subdivide_data

Toggle to TRUE if you want to make comparisons across different levels within your data.

Value

data.frame of differentially expressed genes across comparisons

Details

Aggregate gene expression amongst replicates before performing limma regression