Recluster Filtered Seurat Object.

This template reclusters a filtered Seurat object.

Usage

reclusterFilteredSeuratObject(
  object,
  prepend.txt = "old",
  old.columns.to.save,
  number.of.pcs = 50,
  cluster.resolution.low.range = 0.2,
  cluster.resolution.high.range = 1.2,
  cluster.resolution.range.bins = 0.2,
  reduction.type = "tsne"
)

Arguments

object

The input Seurat Object.

prepend.txt

Text to prepend to old columns to make them unique from new. Default is "old".

old.columns.to.save

Old seurat clustering columns (e.g. SCT_snn_res.0.4) to save.

number.of.pcs

Select the number of principal components for your analysis. Set to 0 to automatically decide. Default is 50.

cluster.resolution.low.range

Select minimum resolution for clustering plots. The lower you set this, the FEWER clusters will be generated. Default is 0.2.

cluster.resolution.high.range

Select maximum resolution for clustering plots. The higher you set this, the MORE clusters will be generated. Default is 1.2.

cluster.resolution.range.bins

Select the bins for your cluster plots. For example, if you input 0.2 as your bin, and have low/high resolution ranges of 0.2 and 0.6, then the template will produce cluster plots at resolutions of 0.2, 0.4 and 0.6. Default is 0.2.

reduction.type

Select the kind of clustering visualization you would like to use to visualize the cell type results ("umap", "tsne", "pca"). Default is "tsne".

Value

Function returns a reclustered Seurat Object with new clustering columns and renamed original clustering columns, along with a plot of the new dimsensionality reduction.

Details

This method reclusters the filtered SO, preserving the original SCT clustering columns with a prepended prefix, and making new SCT clustering columns based on the reclustering. The image returned is the reclustered project.