3D tSNE Coordinate Template from v 75

3D tSNE Coordinate Template from v 75

tSNE_3D_Coordinates(
  Combine_and_Renormalize,
  Gene = c("Gapdh", "GAPDH"),
  Assay = "SCT",
  Max_sample = 10000
)

Arguments

Combine_and_Renormalize

Seurat Object output from previous combine and renormalize template/node.

Gene

Default values: Gapdh, GAPDH. Please enter genes which you would like to visualize. If you don't know what genes to look at, consider consulting your DEG table for interest contrasts.

Assay

Default value: SCT. Select from SCT,RNA,integrated. Select Assay to Plot (default is SCT).

Max_sample

Default value: 10000. Random subsampling of cells without replacement. At the moment only ten thousand cells can be displayed in 3d tSNE viewer.

Value

return the proccessed 3d coordinate dataset