Violin Plot by Metadata

Create violin plot of gene expression data across groups

Usage

violinPlot(
  object,
  assay,
  layer,
  genes,
  group,
  facet_by = "",
  filter_outliers = F,
  outlier_low = 0.05,
  outlier_high = 0.95,
  jitter_points,
  jitter_dot_size
)

Arguments

object

Seurat-class object

assay

Assay to extract gene expression data from (Default: SCT)

layer

Slot to extract gene expression data from (Default: scale.data)

genes

Genes to visualize on the violin plot

group

Split violin plot based on metadata group

facet_by

Split violin plot based on a second metadata group

filter_outliers

Filter outliers from the data (TRUE/FALSE)

outlier_low

Filter lower bound outliers (Default = 0.05)

outlier_high

Filter upper bound outliers (Default = 0.95)

jitter_points

Scatter points on the plot (TRUE/FALSE)

jitter_dot_size

Set size of individual points

Value

violin ggplot2 object

Details

Takes in a list of genes inputted by the user, displays violin plots of genes across groups from a layer-assay with (optional) outliers removed. Can also choose to scale or transform expression data.

Examples

if (FALSE) { # \dontrun{
violinPlot(
  object = seurat,
  assay = "SCT",
  layer = "data",
  genes = c("Cd4", "Cd8a"),
  group = "celltype",
  facet_by = "orig.ident",
  filter_outliers = TRUE,
  jitter_points = TRUE,
  jitter_dot_size = 0.5
)
} # }