Visualize gene expression for provided Genes across your cells as a heatmap

You should see one plot (TSNE or UMAP, your choice) per gene name provided. The intensity of the red color will be relative to the expression of the gene in that cell

colorByGene(
  object,
  samples.to.include,
  gene,
  reduction.type = "umap",
  number.of.rows = 0,
  return.seurat.object = FALSE,
  color = "red",
  point.size = 1,
  point.shape = 16,
  point.transparency = 0.5,
  use.cite.seq.data = FALSE,
  assay = "SCT"
)

Arguments

object

Object of class Seurat

samples.to.include

Samples to be included in the analysis

gene

Genes which you would like to visualize

reduction.type

Select the kind of clustering visualization you would like to use to visualize the cell type results ("umap", "tsne", "pca"). Default is "umap"

number.of.rows

The number of rows you want to arrange your plots into

return.seurat.object

Set to FALSE if you want only a geneset (and not a Seurat object) to be returned

color

The color you want to use in your heatmap (default "red")

point.size

The size of the points representing each cell in your visualization. Default is 1

point.shape

The code for your point shape (R "pch" argument). Default is 16

point.transparency

Set the transparency. Default is 0.5

use.cite.seq.data

TRUE if you would like to plot Antibody clusters from CITEseq instead of scRNA.

assay

Select Assay to Plot (default is SCT)

Value

a Seurat object with additional metadata or gene table and plot

Details

This function must be run downstream of the Sample Names function, as well as be provided a combined Seurat Object such as the one produced by the SingleR Cell Annotation function