Compute ModScore

Returns Seurat-class object with metadata containing ModuleScores and Likely_CellType calls

Usage

modScore(
  object,
  marker.table,
  ms.threshold,
  use_assay = "SCT",
  general.class,
  lvl.vec = c(),
  reduction = "tsne",
  nbins = 10,
  gradient.ft.size = 6,
  violin.ft.size = 6,
  step.size = 0.1
)

Arguments

object

Seurat-class object

marker.table

Table of marker genes for each celltype (column names of the table), append "_prot" or "_neg" for proteins or negative markers

general.class

Base population of cells to classify

reduction

Choose among tsne, umap, and pca (Default: tsne)

nbins

Number of bins for storing control features and analyzing average expression (Default: 10)

gradient.ft.size

Set size of axis labels on gradient density plot of ModuleScore distribution (Default: 6)

violin.ft.size

Set size of axis labels on violin plot of ModuleScore distribution (Default: 6)

step.size

Set step size of distribution plots (Default: 0.1)

cite.seq

Set to TRUE if there are CITE-seq markers in marker.table (Default: FALSE)

celltypes

Vector of celltypes from marker.table to screen for

threshold

Specify bimodal thresholds for cell classification, should be of the same length as celltypes vector

multi.lvl

Toggle to TRUE if there are subpopulations of cells you want to screen for (Default: FALSE)

lvl.df

Table of subpopulation levels and parent-child information (e.g. Tcells-CD4, Tcells-CD8)

Value

List containing annotated dimension plot with ModuleScore distribution of cell marker gene, Seurat Object with cell classification metadata

Details

Analyzed features are binned based on averaged expression; control features are randomly selected from each bin.